+++ cmd/root.go
@@ -85,11 +85,13 @@ func init() {
 	rootCmd.Flags().StringP("log-file", "l", "", "path to cdk log file")
 	rootCmd.Flags().StringP("tag-id", "t", "stack", "unique identifier for stack within pipeline")
 	rootCmd.Flags().BoolP("delete", "d", true, "delete comments when no changes are detected for a specific tag id")
-	rootCmd.Flags().String("vcs", "github", "Version Control System [github|bitbucket|gitlab]")
+	rootCmd.Flags().String("vcs", "github", "Version Control System [github|github-enterprise|bitbucket|gitlab]")
 	rootCmd.Flags().String("ci", "circleci", "CI System used [circleci|bitbucket|gitlab]")
 	rootCmd.Flags().StringP("user", "u", "", "Optional set username for token (required for bitbucket)")
 	rootCmd.Flags().String("gitlab-url", "https://gitlab.com/", "Optional set gitlab url")
+	rootCmd.Flags().String("github-host", "", "Optional set host for GitHub Enterprise")
 	rootCmd.Flags().Bool("no-post-mode", false, "Optional do not post comment to VCS, instead write additional file and print diff to stdout")
+	rootCmd.Flags().Bool("disable-collapse", false, "Collapsible comments are enabled by default for GitHub and GitLab. When set to true it will not use collapsed sections.")
 
 	// mapping for viper [mapstruct value, flag name]
 	viperMappings := make(map[string]string)
@@ -102,9 +104,11 @@ func init() {
 	viperMappings["TAG_ID"] = "tag-id"
 	viperMappings["DELETE_COMMENT"] = "delete"
 	viperMappings["NO_POST_MODE"] = "no-post-mode"
+	viperMappings["DISABLE_COLLAPSE"] = "disable-collapse"
 	viperMappings["VERSION_CONTROL_SYSTEM"] = "vcs"
 	viperMappings["CI_SYSTEM"] = "ci"
 	viperMappings["URL"] = "gitlab-url"
+	viperMappings["GITHUB_ENTERPRISE_HOST"] = "github-host"
 
 	for k, v := range viperMappings {
 		err := viper.BindPFlag(k, rootCmd.Flags().Lookup(v))