+++ config/app_test.go
@@ -11,12 +11,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-type testCase struct {
-	input    string
-	expected int
-	err      error
-}
-
 type testCaseInit struct {
 	description    string
 	inputConfig    NotifierConfig
@@ -29,40 +23,82 @@ type testCaseInit struct {
 
 func TestNotifierConfig_ReadPullRequestFromEnv(t *testing.T) {
 	logrus.SetLevel(7)
-	testCases := []testCase{
+	tests := []struct {
+		name     string
+		inputUrl string
+		want     *PullRequest
+		wantErr  bool
+	}{
 		{
-			input:    "https://github.com/pansenentertainment/uepsilon/pull/1361",
-			expected: 1361,
-			err:      nil,
+			name:     "Valid PR URL",
+			inputUrl: "https://github.com/owner/repo/pull/123",
+			want: &PullRequest{
+				Host:   "github.com",
+				Owner:  "owner",
+				Repo:   "repo",
+				Number: 123,
+			},
+			wantErr: false,
 		},
 		{
-			input:    "/1361",
-			expected: 1361,
-			err:      nil,
+			name:     "Valid PR URL with custom host",
+			inputUrl: "https://mycompany.com/owner/repo/pull/123",
+			want: &PullRequest{
+				Host:   "mycompany.com",
+				Owner:  "owner",
+				Repo:   "repo",
+				Number: 123,
+			},
+			wantErr: false,
 		},
 		{
-			input:    "1361",
-			expected: 1361,
-			err:      nil,
+			name:     "Valid PR number",
+			inputUrl: "9",
+			want: &PullRequest{
+				Host:   "",
+				Owner:  "",
+				Repo:   "",
+				Number: 9,
+			},
+			wantErr: false,
 		},
 		{
-			// not setting pull request id should not throw error
-			input:    "",
-			expected: 0,
-			err:      nil,
+			name:     "Invalid PR URL",
+			inputUrl: "",
+			want: &PullRequest{
+				Number: 0,
+			},
+			wantErr: false,
 		},
 		{
-			input:    "https://github.com/pansenentertainment/uepsilon/pull",
-			expected: 0,
-			err:      &strconv.NumError{},
+			name:     "Invalid PR URL with null",
+			inputUrl: "https://github.com/owner/repo/pull/null",
+			want:     nil,
+			wantErr:  true,
+		},
+		{
+			name:     "Invalid URL Scheme",
+			inputUrl: "::::",
+			want:     nil,
+			wantErr:  true,
 		},
 	}
 
-	for _, c := range testCases {
-		_ = os.Setenv(EnvCiCircleCiPullRequestID, c.input)
-		actual, err := readPullRequestFromEnv()
-		assert.IsType(t, c.err, err)
-		assert.Equal(t, c.expected, actual)
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			pr := &PullRequest{}
+			err := pr.ConvertUrlToPullRequest(tt.inputUrl)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ConvertUrlToPullRequest() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if tt.wantErr {
+				return
+			}
+			if *pr != *tt.want {
+				t.Errorf("ConvertUrlToPullRequest() = %+v, want %+v", pr, tt.want)
+			}
+		})
 	}
 }
 
@@ -88,6 +124,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "some-token",
 				PullRequestID: 23,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
 			err: nil,
 		},
@@ -112,6 +149,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "some-token",
 				PullRequestID: 0,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
 			err: nil,
 		},
@@ -136,12 +174,13 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "some-token",
 				PullRequestID: 0,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
-			err: &strconv.NumError{
-				Func: "ParseInt",
+			err: fmt.Errorf("Unable to extract pull request number from url '%s': %w", "23as", &strconv.NumError{
+				Func: "Atoi",
 				Num:  "23as",
 				Err:  strconv.ErrSyntax,
-			},
+			}),
 		},
 		{
 			description: "test bitbucket ci config override",
@@ -167,6 +206,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "bitbucket-token",
 				PullRequestID: 12,
 				Ci:            CiBitbucket,
+				Vcs: 	  VcsGithub,
 			},
 			err: nil,
 		},
@@ -193,6 +233,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "bitbucket-token",
 				PullRequestID: 12,
 				Ci:            "",
+				Vcs: 	  VcsGithub,
 			},
 			err: nil,
 		},
@@ -217,6 +258,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "",
 				PullRequestID: 23,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
 			err: &ValidationError{"token", []string{"TOKEN", EnvGithubToken, EnvBitbucketToken, EnvGitlabToken}},
 		},
@@ -241,6 +283,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "some-token",
 				PullRequestID: 23,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
 			err: &ValidationError{"repo", []string{"REPO_NAME", EnvCiCircleCiRepoName, EnvCiBitbucketRepoName, EnvCiGitlabRepoName}},
 		},
@@ -265,6 +308,7 @@ func TestNotifierConfig_Init(t *testing.T) {
 				Token:         "some-token",
 				PullRequestID: 23,
 				Ci:            CiCircleCi,
+				Vcs: 	  VcsGithub,
 			},
 			err: &ValidationError{"owner", []string{"REPO_OWNER", EnvCiCircleCiRepoOwner, EnvCiBitbucketRepoOwner, EnvCiGitlabRepoOwner}},
 		},
@@ -283,23 +327,53 @@ func TestNotifierConfig_Init(t *testing.T) {
 				TagID:      "no-post",
 				NoPostMode: true,
 				Ci:         CiCircleCi,
+				Vcs: 	  VcsGithub,
+			},
+			err: nil,
+		},
+		{
+			description: "test github enterprise host is set from circleci env",
+			inputConfig: NotifierConfig{
+				LogFile: "./cdk.log",
+				RepoName:      "repo",
+				RepoOwner:     "owner",
+			},
+			envVars: map[string]string{
+				EnvCiCircleCiPullRequestID: "https://github.your-company.com/owner/repo/pull/1",
+				EnvGithubToken:             "some-token",
+			},
+			ci:  CiCircleCi,
+			vcs: VcsGithubEnterprise,
+			expectedConfig: NotifierConfig{
+				LogFile: "./cdk.log",
+				GithubHost: "github.your-company.com",
+				Ci: CiCircleCi,
+				PullRequestID: 1,
+				Vcs: VcsGithubEnterprise,
+				RepoName:      "repo",
+				RepoOwner:     "owner",
+				Token:         "some-token",
 			},
 			err: nil,
 		},
 	}
 	for _, c := range testCasesInit {
-		t.Log(c.description)
-		for k, v := range c.envVars {
-			fmt.Printf("set env variable %s: %s\n", k, v)
-			_ = os.Setenv(k, v)
-		}
-		viper.Set("ci_system", c.ci)
-		viper.Set("vcs", c.vcs)
-		err := c.inputConfig.Init()
-		assert.Equal(t, c.err, err, c.description)
-		assert.Equal(t, c.expectedConfig, c.inputConfig, c.description)
-		for k := range c.envVars {
-			_ = os.Unsetenv(k)
-		}
+		t.Run(c.description, func(t *testing.T) {
+			t.Log(c.description)
+			for k, v := range c.envVars {
+				fmt.Printf("set env variable %s: %s\n", k, v)
+				_ = os.Setenv(k, v)
+			}
+			viper.Set("ci_system", c.ci)
+			viper.Set("version_control_system", c.vcs)
+			err := c.inputConfig.Init()
+			assert.Equal(t, c.err, err, c.description)
+			assert.Equal(t, c.expectedConfig, c.inputConfig, c.description)
+			defer func() {
+				for k := range c.envVars {
+				_ = os.Unsetenv(k)
+				}
+			}()
+		})
 	}
 }